+++ openprocurement/auctions/swiftsure/tests/chronograph.py
@@ -43,9 +43,6 @@ from openprocurement.auctions.swiftsure.tests import fixtures
 from openprocurement.auctions.swiftsure.tests.blanks.chronograph_blanks import (
     # AuctionSwitchQualificationResourceTest
     switch_to_qualification,
-    # AuctionAuctionPeriodResourceTest
-    set_auction_period,
-    reset_auction_period
 )
 
 
@@ -81,13 +78,6 @@ class AuctionLotSwitchAuctionResourceTest(AuctionSwitchAuctionResourceTest):
 class AuctionLotSwitchUnsuccessfulResourceTest(
         AuctionSwitchUnsuccessfulResourceTest):
     initial_lots = test_lots
-
-
-class AuctionAuctionPeriodResourceTest(BaseAuctionWebTest):
-    initial_bids = test_bids
-
-    test_set_auction_period = snitch(set_auction_period)
-    test_reset_auction_period = snitch(reset_auction_period)
 
 
 class AuctionAward2BidsSwitchResourceTest(
@@ -190,11 +180,6 @@ class AuctionAwardSwitch2ResourceTest(BaseAuctionWebTest):
         switch_verification_to_unsuccessful_2)
     test_switch_active_to_unsuccessful_2 = snitch(
         switch_active_to_unsuccessful_2)
-
-
-@unittest.skip("option not available")
-class AuctionLotAuctionPeriodResourceTest(AuctionAuctionPeriodResourceTest):
-    initial_lots = test_lots
 
 
 @unittest.skip("option not available")
@@ -315,10 +300,8 @@ def suite():
         AuctionLotSwitchQualificationResourceTest))
     tests.addTest(unittest.makeSuite(AuctionLotSwitchAuctionResourceTest))
     tests.addTest(unittest.makeSuite(AuctionLotSwitchUnsuccessfulResourceTest))
-    tests.addTest(unittest.makeSuite(AuctionAuctionPeriodResourceTest))
     tests.addTest(unittest.makeSuite(AuctionAward2BidsSwitchResourceTest))
     tests.addTest(unittest.makeSuite(AuctionAwardSwitch2ResourceTest))
-    tests.addTest(unittest.makeSuite(AuctionLotAuctionPeriodResourceTest))
     tests.addTest(unittest.makeSuite(AuctionComplaintSwitchResourceTest))
     tests.addTest(unittest.makeSuite(AuctionLotComplaintSwitchResourceTest))
     tests.addTest(unittest.makeSuite(AuctionAwardComplaintSwitchResourceTest))
